Move password to landing page
