+++ src/scenes/ConnectWallet/ConnectWallet.tsx
@@ -1,0 +1,7 @@
+import { RouteComponentProps } from '@reach/router';
+
+function ConnectWallet(_: RouteComponentProps) {
+  return <div> Wallet Connect</div>;
+}
+
+export default ConnectWallet;
